extract render items to own element
